pyproject.toml
[project]
name = "fast-api-template"
version = "0.1.0"
description = ""
authors = [
    {name = "Mateus Gurgel",email = "mateusgurgel15@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "dependency-injector (>=4.47.1,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "stripe (>=13.1.0,<14.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "scripts.dev:run"
create_tables = "scripts.create_tables:run"
create_subscription_plan = "scripts.create_subscription_plan:app"


[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
pre-commit = "^4.2.0"

[tool.mypy]
python_version = "3.10"
strict = true
exclude = "(venv|env|\\.venv|tests/.*_test\\.py)"
check_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
